+++ src/test/java/io/justedlev/msrv/ssr/SimpleEurekaServerApplicationTests.java
@@ -1,4 +1,4 @@
-package dev.justedlev.eurekaserver;
+package io.justedlev.msrv.ssr;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
